+++ nnunet/training/network_training/nnUNetTrainerV2_CascadeFullRes.py
@@ -287,7 +287,7 @@ class nnUNetTrainerV2CascadeFullRes(nnUNetTrainerV2):
                 patching system python code. We circumvent that problem here by saving softmax_pred to a npy file that will 
                 then be read (and finally deleted) by the Process. save_segmentation_nifti_from_softmax can take either 
                 filename or np.ndarray and will handle this automatically"""
-                if np.prod(softmax_pred.shape) > (2e9 / 4 * 0.85):  # *0.85 just to be save
+                if np.prod(softmax_pred.shape, dtype=np.double) > (2e9 / 4 * 0.85):  # *0.85 just to be save
                     np.save(join(output_folder, fname + ".npy"), softmax_pred)
                     softmax_pred = join(output_folder, fname + ".npy")
